software/pc/requirements.txt
intelhex==2.3.0
pyserial==3.5
pywin32; sys_platform == "win32" 